# api/openapi.yaml
openapi: 3.1.0
info:
  title: Violet Deterministic API
  version: 0.4.0
servers:
  - url: http://localhost:4020
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
...
paths:
...
  /v1/tools:
    get:
      summary: List API tools and CLI command mappings for agent orchestration
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tool catalog
...
  /v1/studio/jobs/{id}/run:
    post:
      summary: Execute a named run target (web/mobile/api/verify/all) for generated artifacts and runtime smoke checks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Run target result
...
  /v1/studio/jobs/{id}/runtime/{client}/{asset}:
    get:
      summary: Serve generated runtime assets for web/mobile preview clients
      description: Used by preview iframes to load executable CSS/JS for the generated app experience.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: client
          in: path
          required: true
          schema:
            type: string
            enum: [web, mobile]
        - name: asset
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Runtime asset bytes
...